% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{calc_dims}
\alias{calc_dims}
\title{Calculate reduction optimal dimensions}
\usage{
calc_dims(seurat_object, reduction = "pca")
}
\arguments{
\item{seurat_object}{Seurat object to use.}

\item{reduction}{Name of the reduction to use. Default to "pca".}
}
\value{
A Seurat object.
}
\description{
This function calculates the optimal dimensions to use for UMAP, tSNE, clustering etc starting from the values of a
pre-calculated dimensional reduction (e.g. pca).
}
\details{
This function will first calculate stdev associated with each embedding of a reduction, if not already calculated.
It then retrieve the \% of variation associated with each PC, as well as the comulative sum; and finally it prints
both the dimension that exhibits cumulative percent greater than 90\% and \% variation associated with the PC as less than 5 (pc90) and the
first dimension that exhibits a difference with the next < 0.1 (pc01). Those two values are then stored in the \@misc slot of the
corresponding reduction.
}
\examples{
data("pbmc_small")

# Specify the reduction, the axis prefix and the group_by variable
pbmc_small <- calc_dims(pbmc_small)

}
\concept{DimRed}
