% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{calculate_qc_mad_outliers}
\alias{calculate_qc_mad_outliers}
\title{Calculate QC metrics outliers with MADs}
\usage{
calculate_qc_mad_outliers(
  seurat_object,
  assay = NULL,
  batch = "orig.ident",
  extra = list()
)
}
\arguments{
\item{seurat_object}{Seurat object to use.}

\item{assay}{Assay to use to select nCount and nFeature meta.data. By default it takes the current active assay.}

\item{batch}{Name of the column to pass to \code{isOutlier(batch)}. Default to "orig.ident".}

\item{extra}{Optional list containing custom metrics to calculate or default values to override.}
}
\value{
A Seurat object.
}
\description{
This function calculates, for each qc metric previously calculated with \code{\link[=calculate_qc_metrics]{calculate_qc_metrics()}} using the MAD approach
(check \code{\link[scater:reexports]{scater::isOutlier()}} for more info).
By default, nCount and nFeature for the current assay, as well as mitoRatio and riboRatio are examinated. To see default values,
run \code{\link[=view_default_qc_mads]{view_default_qc_mads()}}.
For each metric, it adds a new column in meta.data called metric_outlier which mark whether a cell is an outlier (TRUE) for that metric or not (FALSE);
it also creates a dataframe in seurat_object@misc$qc_thresholds containing the calculated threshold values.
}
\section{Add custom metrics}{

To add custom metrics, provide a list as argument to \code{extra} in this format:
list("name_of_meta.data_column" = list("nmads" = c(lower_value, upper_value), "log" = boolean)). Both nmads and log are REQUIRED and cannot be omitted.
lower_value or upper_value can be set to \code{NA} if that threshold should not be calculated.
}

\section{Override/skip default values}{

To override default values, provide the corresponding list in extra argument as if it is an extra column to calculate. It will override default parameters for that metric.
To skip a default metric, set it to NULL as metric element list (e.g \code{extra = list("mitoRatio" = NULL)})
}

\examples{
data("pbmc_small")
pbmc_small <- calculate_qc_metrics(pbmc_small)

# Default with no customization
pbmc_small <- calculate_qc_mad_outliers(pbmc_small)

# Setting batch and omit riboRatio calculation
pbmc_small <- calculate_qc_mad_outliers(pbmc_small,
                                        batch = "groups",
                                        extra = list("riboRatio" = NULL))


}
\concept{QC}
