% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{counts}
\alias{counts}
\title{Gene expression matrix}
\format{
\subsection{\code{counts}}{

Sparse dgc matrix of gene expression by cells
}
}
\usage{
counts
}
\description{
An example subset of  1\% of the original data 2623 cells and 30 genes are
included with the package. To run the package with the full data, overwrite
metadata and counts objects in the environment.
}
\keyword{datasets}
