% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{create_parameter_list}
\alias{create_parameter_list}
\title{Create parameter list for MAD outlier}
\usage{
create_parameter_list(assay = "RNA", extra = list())
}
\arguments{
\item{assay}{Assay to use to select nCount and nFeature meta.data. By default it takes the current active assay.}

\item{extra}{Optional list containing custom metrics to calculate or default values to override.}
}
\value{
A list with elemets like this: "name_of_meta.data_column" = list("nmads" = c(lower_value, upper_value), "log" = boolean).
}
\description{
This function creates a list with the parameter to use for MAD outlier QC calculation.
}
\concept{QC}
