% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap_expr.R
\name{extract_gene}
\alias{extract_gene}
\title{extracts gene expression data into metadata}
\usage{
extract_gene(metadata, expr, genes)
}
\arguments{
\item{metadata}{cell metadata}

\item{expr}{count matrix, cells in columns, genes in rows}

\item{genes}{character vector of gene names}
}
\value{
tibble
}
\description{
extracts gene expression data into metadata
}
