% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{genelists}
\alias{genelists}
\title{genelists}
\format{
\subsection{\code{genelists}}{

A list containing 5 sets of genes
}
}
\source{
Smillie 2019 methods, Ferarro 2014 supplemental
}
\usage{
genelists
}
\description{
Sets of genes that are interesting
}
\details{
Smillie 2019 has sets of genes that were able to separate epithelial,
stromal, and immune cells

Ferraro 2014 has sets of genes that are differentially expressed in Tregs vs
Tconv cells.
}
\keyword{datasets}
