% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{metadata}
\alias{metadata}
\title{Cell metadata}
\format{
\subsection{\code{metadata}}{

tibble of cell metadata.
}
}
\usage{
metadata
}
\description{
An example subset of 1\% of the original data 2623 cells are included with
the package. To run the package with the full data, overwrite metadata and
counts objects in the environment.
}
\keyword{datasets}
