% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot_clusters_metadata}
\alias{plot_clusters_metadata}
\title{Plot clusters metadata}
\usage{
plot_clusters_metadata(
  seurat_object,
  reduction = "pca",
  group_by = "seurat_clusters",
  split_by = "orig.ident",
  ncol = 3,
  dims = 1:2,
  cols_vector = NULL,
  axis_prefix = "PC_",
  transpose = FALSE,
  heights = c(2, 1, 1),
  return_list = FALSE
)
}
\arguments{
\item{seurat_object}{Seurat object to use.}

\item{reduction}{Name of the reduction to use. Default to "pca".}

\item{group_by}{Name of the column to use to create groups of samples. Default to "seurat_clusters".}

\item{split_by}{Name of the column to use to split the plot for. Default to "orig.ident".}

\item{ncol}{Number of column to pass to ggarrange(ncol) when arranging the different dimplots.}

\item{dims}{2-values vector with the dimension to use for plotting. Default to 1:2.}

\item{cols_vector}{Optional named vector that specifies the colors to use for each value of the group_by variable.
Default to NULL, using turbo as default palette.}

\item{axis_prefix}{Prefix to add to dimensions as axis dimension names. Default to "PC_".}

\item{transpose}{Whether to transpose table to better fit the image. Default to FALSE.}

\item{heights}{2-values vector to pass to ggarrange(height). It sets the height proportion between the main DimPlot and the
panel of the splitted dimplots. Default to c(2, 1, 1).}

\item{return_list}{Whether to return a list of ggplot objects (TRUE) or a ggarrange object with all the plots (FALSE). Default to FALSE.}
}
\value{
A list of ggplot objects or a ggarrange object.
}
\description{
This function creates splitted dimplots based on a meta.data categorical column, a table of the number and
percentages of cells per cluster (or category of a categorical meta.data columns) for each category of the meta.data column.
}
\examples{
data("pbmc_small")


# Selecting reduction and group_by column
plot_clusters_metadata(pbmc_small, reduction = "tsne",
                       group_by = "letter.idents", split_by = "groups",
                       axis_prefix = "TSNE_")

}
\concept{Clustering}
