% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.R
\name{plot_dotplot}
\alias{plot_dotplot}
\alias{plot_dotplot_groups}
\title{plots a heatmap / dotplot}
\usage{
plot_dotplot(
  metadata,
  expr,
  genes,
  zscore = TRUE,
  zscore_max = 3,
  zscore_min = 0,
  celltype_col,
  cellid_col = "cellid",
  fontsize_p = 1,
  facet_row = NULL
)

plot_dotplot_groups(
  metadata,
  expr,
  genes,
  zscore = TRUE,
  zscore_max = 3,
  zscore_min = 0,
  celltype_col,
  cellid_col = "cellid",
  fontsize_p = 1,
  facet_row = NULL,
  groups
)
}
\arguments{
\item{metadata}{cells metadata}

\item{expr}{counts matrix}

\item{genes}{character vector of gene names}

\item{zscore}{zscores expression by row (gene)}

\item{zscore_max}{zscores over this will have same color. Only when zscore =
TRUE.}

\item{zscore_min}{zscores below this will be white.}

\item{celltype_col}{column in metadata that contains celltype cluster labels}

\item{cellid_col}{column in metadata that contains cell IDs, matches rownames
of expr. expects a character}

\item{fontsize_p}{a multiplier that controls fontsize.}

\item{facet_row}{variable to facet rows (cells) by}

\item{groups}{column of metadata to plot on x-axis (e.g. disease)}
}
\value{
ggplot2
}
\description{
plots a heatmap / dotplot
}
\section{Functions}{
\itemize{
\item \code{plot_dotplot_groups()}: puts sample groups on the x-axis and column facets
by gene

}}
