% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genebycell.R
\name{plot_genebycell}
\alias{plot_genebycell}
\title{plots expression by cell cluster}
\usage{
plot_genebycell(
  metadata,
  cellid_col = "cellid",
  expr,
  genes,
  celltype_col,
  type = "boxplot",
  zscore = FALSE,
  add_jitter = FALSE,
  jitterheight = 0.75,
  alpha = 0.5,
  pointsize = 1,
  expand_mult = 0,
  expand_add = 0.45,
  fontsize_p = 1,
  p_expr_highlight_thresh = 0.05,
  pexpr_label_size = 2.25,
  title = ""
)
}
\arguments{
\item{metadata}{cell metadata}

\item{cellid_col}{column in metadata containing cell identifiers, should
match rownames(counts). Default is "cellid"}

\item{expr}{count matrix, cells in columns, genes in rows}

\item{genes}{character vector of gene names}

\item{celltype_col}{column in metadata referring to cell type identity}

\item{type}{one of "mean_se", "bar_mean", "bar_mean_se", "bar_median",
"violin", "boxplot"}

\item{zscore}{Whether expression should be zscore or not}

\item{add_jitter}{adds jittered points default = FALSE,}

\item{jitterheight}{multiplier that controls jitter height. Default = 0.75}

\item{alpha}{controls the alpha of jittered points = 0.5}

\item{pointsize}{= 1,}

\item{expand_mult}{expand_mult and *_add increases limit of x axis to make
space for text labels. Default mult is 0}

\item{expand_add}{default is 0.45}

\item{fontsize_p}{Multiplier that controls relative fontsize. Default is 1}

\item{p_expr_highlight_thresh}{clusters over/under this proportion (0-1) are
highlighted}

\item{pexpr_label_size}{controls size of percent expression labels. Default
is 2.25}

\item{title}{title of plot}
}
\value{
ggplot2
}
\description{
Use extract_gene so that gene expression exists as columns of metadata.
}
