% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_qc_metrics}
\alias{plot_qc_metrics}
\title{Create all density plots for a QC metrics}
\usage{
plot_qc_metrics(
  seurat_object,
  assay = NULL,
  fill = "orig.ident",
  fill_vector = NULL,
  alpha = 0.2,
  count_thresh = 1000,
  feature_thresh = 1000,
  log10GenesPerUMI_thresh = 0.8,
  mito_thresh = 0.15,
  ribo_thresh = 0.05,
  hemo_thresh = NULL
)
}
\arguments{
\item{seurat_object}{Seurat object to use to create the plot.}

\item{assay}{Assay to use to select nCount and nFeature meta.data. By default it takes the current active assay.}

\item{fill}{Name of the column to use to pass to aes(fill) in geom_density.
It is used to split the data into groups and draw a density plot for each value of it. Default to "orig.ident".}

\item{fill_vector}{Optional named vector used to assign fill colors to the different group of fill variable.
Default to NULL.}

\item{alpha}{Value to pass to alpha in geom_density. Default to 0.2.}

\item{count_thresh}{Value of the count threshold line to pass to qc_density_plot(threshold). Default to 1000.}

\item{feature_thresh}{Value of the feature threshold line to pass to qc_density_plot(threshold). Default to 1000.}

\item{log10GenesPerUMI_thresh}{Value of the log10GenesPerUMI threshold line to pass to qc_density_plot(threshold). Default to 0.8}

\item{mito_thresh}{Value of the mitoRatio threshold line to pass to qc_density_plot(threshold). Default to 0.15.}

\item{ribo_thresh}{Value of the riboRatio threshold line to pass to qc_density_plot(threshold). Default to 0.05.}

\item{hemo_thresh}{Value of the hemoRatio threshold line to pass to qc_density_plot(threshold). Default to 0.NULL.}
}
\value{
A ggarrange object with the plots.
}
\description{
This function is a wrapper around qc_density_plot. It creates density plots for different QC metrics
(nCount, nFeature, log10GenesPerUMI, mitochondrial gene ratio, ribosomal gene ratio and hemoglobin gene ratio) and wrap them up
into a single plot.
}
\details{
For nCount and nFeature density plots, the default x_trans is set to log10.
}
\examples{
data("pbmc_small")
pbmc_small <- calculate_qc_metrics(pbmc_small)

# Basic plot
plot_qc_metrics(pbmc_small)

# Specify other thresholds
plot_qc_metrics(pbmc_small, count_thresh = 750,
                feature_thresh = 900,
                mito_thresh = 0.2)

}
\concept{QC}
