% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_qc_metrics_outliers}
\alias{plot_qc_metrics_outliers}
\title{Plot qc metrics outliers}
\usage{
plot_qc_metrics_outliers(
  seurat_object,
  split_by = "orig.ident",
  fill_vector = NULL,
  metrics = c(nCount_RNA = "log10", nFeature_RNA = "log10", mitoRatio = "identity",
    riboRatio = "identity"),
  return_list = FALSE
)
}
\arguments{
\item{seurat_object}{Seurat object to use.}

\item{split_by}{Name of the column used to group_by the violins (usually should correspond to the one used as \code{batch} in \code{\link[=calculate_qc_mad_outliers]{calculate_qc_mad_outliers()}}).
Default to "orig.ident".}

\item{fill_vector}{Optional named vector used to assign fill colors to the different group of fill variable.
Default to NULL.}

\item{metrics}{Named vector with metrics to plot as names and the desiretd x_trans as value.
Default to metrics = c("nCount_RNA" = "log10", "nFeature_RNA" = "log10", "mitoRatio" = "identity", "riboRatio" = "identity")}

\item{return_list}{Whether to return a list of ggplot objects (TRUE) or a ggarrange object with all the plots (FALSE). Default to FALSE.}
}
\value{
A list of ggplot objects or a ggarrange object.
}
\description{
This function is a wrapper around \code{\link[=plot_violin_outliers]{plot_violin_outliers()}}. It creates a list of violin plots for the desired qc metrics (or meta.data column).
}
\examples{
data("pbmc_small")
pbmc_small <- calculate_qc_metrics(pbmc_small)
pbmc_small <- calculate_qc_mad_outliers(pbmc_small)

# Basic plot
plot_qc_metrics_outliers(pbmc_small)

# Split violin based on a column
pbmc_small <- calculate_qc_mad_outliers(pbmc_small, batch = "groups")
plot_qc_metrics_outliers(pbmc_small, split_by = "groups")

}
\concept{QC}
