% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot_reduction_res}
\alias{plot_reduction_res}
\title{Plot reduction at different resolutions}
\usage{
plot_reduction_res(
  seurat_object,
  reduction = "pca",
  column_pattern = "RNA_snn_res",
  label_size = 6,
  axis_prefix = "PC_",
  return_list = FALSE
)
}
\arguments{
\item{seurat_object}{Seurat object to use.}

\item{reduction}{Name of the reduction to use. Default to "pca".}

\item{column_pattern}{Pattern used to select meta.data column to plot. Default to "RNA_snn_res".}

\item{label_size}{Value to pass to DimPlot(label_size). Default to 6.}

\item{axis_prefix}{Prefix to add to "1" and "2" as axis dimension names. Default to "PC_".}

\item{return_list}{Whether to return a list of ggplot objects (TRUE) or a ggarrange object with all the plots (FALSE). Default to FALSE.}
}
\value{
A list of ggplot objects or a ggarrange object.
}
\description{
This function is used to plot different clustering resolutions on a reduction.
}
\examples{
data("pbmc_small")

# With default parameters
plot_reduction_res(pbmc_small)

# Selecting different reduction
plot_reduction_res(pbmc_small, reduction = "tsne")

}
\concept{Clustering}
