% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot_splitted_clusters}
\alias{plot_splitted_clusters}
\title{Plot splitted highlighted clusters}
\usage{
plot_splitted_clusters(
  seurat_object,
  reduction = "pca",
  group_by = "orig.ident",
  dims = 1:2,
  cols_vector = NULL,
  axis_prefix = "PC_",
  heights = c(1, 2),
  return_list = FALSE
)
}
\arguments{
\item{seurat_object}{Seurat object to use.}

\item{reduction}{Name of the reduction to use. Default to "pca".}

\item{group_by}{Name of the column to use to create groups of samples. Default to "orig.ident".}

\item{dims}{2-values vector with the dimension to use for plotting. Default to 1:2.}

\item{cols_vector}{Optional named vector that specifies the colors to use for each value of the group_by variable.
Default to NULL, using turbo as default palette.}

\item{axis_prefix}{Prefix to add to dimensions as axis dimension names. Default to "PC_".}

\item{heights}{2-values vector to pass to ggarrange(height). It sets the height proportion between the main DimPlot and the
panel of the splitted dimplots. Default to c(1, 2).}

\item{return_list}{Whether to return a list of ggplot objects (TRUE) or a ggarrange object with all the plots (FALSE). Default to FALSE.}
}
\value{
A list of ggplot objects or a ggarrange object.
}
\description{
This function generate plots to highlight the position of each cluster (or meta.data categorical column) in the reduction space.
It create a main plot (a classic DimPlot), along with a subplot for each cluster (or category), highlighting the position of
the samples belonging to it in the reduction space.
}
\examples{
data("pbmc_small")

# With default parameters
plot_splitted_clusters(pbmc_small)

# Selecting different reduction and group_by column
plot_splitted_clusters(pbmc_small, reduction = "tsne",
                       group_by = "letter.idents",
                       axis_prefix = "TSNE_")

}
\concept{Clustering}
