% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{plot_splitted_dimplots}
\alias{plot_splitted_dimplots}
\title{Plot splitted dimplots}
\usage{
plot_splitted_dimplots(
  seurat_object,
  reduction = "pca",
  group_by = "orig.ident",
  dims = 1:2,
  cols_vector = NULL,
  heights = c(1, 2),
  axis_prefix = "PCA"
)
}
\arguments{
\item{seurat_object}{Seurat object to use.}

\item{reduction}{Name of the reduction to use for plotting. Default to "pca".}

\item{group_by}{Name of the variable to use for grouping cells into dimplots. Default to "orig.ident".}

\item{dims}{2-values vector with the dimension to use for plotting. Default to 1:2.}

\item{cols_vector}{Optional named vector that specifies the colors to use for each value of the group_by variable.
Default to NULL, using viridis as default palette.}

\item{heights}{2-values vector to pass to ggarrange(height). It sets the height proportion between the main DimPlot and the
panel of the splitted dimplots. Default to c(1, 2).}

\item{axis_prefix}{Prefix to add to dims to label the axes. Default to "PCA".}
}
\value{
A ggarrange object.
}
\description{
This function creates a collection of dimplots: one generic grouped by the variable
set, and one for each unique value of the variable set, arranging them into a single plot.
}
\details{
Each subplot (one for each category of the group_by variable) will have the dots of the corresponding category colored
as specified in cols_vector (if specified) and all the other dots colored in gray, in order to see how the samples of that
category spread throughout the DimPlot compared to all the others.
}
\examples{
data("pbmc_small")
plot_splitted_dimplots(pbmc_small)

# Specify the reduction, the axis prefix and the group_by variable
plot_splitted_dimplots(pbmc_small,
                       reduction = "tsne",
                       axis_prefix = "tsne",
                       group_by = "RNA_snn_res.1")

}
\concept{DimRed}
