% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_top_n_genes}
\alias{plot_top_n_genes}
\title{Plot top expressed features.}
\usage{
plot_top_n_genes(seurat_object, assay = "RNA", slot = "count", n = 10)
}
\arguments{
\item{seurat_object}{Seurat object to use to create the plot.}

\item{assay}{Name of the assay to use for extrapolating the data. Default to "RNA".}

\item{slot}{Name of the slot to use. Default to "counts".}

\item{n}{Number of top features to count. Default to 10.}
}
\value{
A ggplot list object with the plot.
}
\description{
This function creates a boxplot representing the top n features, based on median value of \% of expression in cells.
}
\examples{
data("pbmc_small")

# Basic plot
plot_top_n_genes(pbmc_small)

# Specify n
plot_top_n_genes(pbmc_small, n = 25)

}
\concept{QC}
