% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap_expr.R
\name{plot_umap_expr}
\alias{plot_umap_expr}
\title{plots expression on umap coordinates}
\usage{
plot_umap_expr(
  metadata,
  genes,
  UMAP1 = "UMAP_1",
  UMAP2 = "UMAP_2",
  zscore = TRUE,
  pointsize = 1.2,
  pixels = 712,
  alpha = 1,
  breaks = 4,
  ncol = NULL,
  nrow = NULL
)
}
\arguments{
\item{metadata}{cells metadata}

\item{genes}{character vector of gene names}

\item{zscore}{Whether expression should be zscore or not}

\item{pointsize}{non-integers work best}

\item{pixels}{resolution for ggscattermore}
}
\value{
ggplot2
}
\description{
Use extract_gene so that gene expression exists as columns of metadata.
}
\details{
Color is scaled from the minimum to 99th percentile only if the difference in
99th percentile and max is greater than 1 unit, and the 99th percentile isn't
equal to 0. This addresses the situation where a very small percentage (<1\%)
cells have massively higher expression of a gene (common for small clusters),
which makes it hard to visualize the variability of the other 99\% of the
data.
}
