% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{plot_variable_features}
\alias{plot_variable_features}
\title{Plot variable features}
\usage{
plot_variable_features(seurat_object, top_n = 20, cols = c("black", "blue"))
}
\arguments{
\item{seurat_object}{Seurat object to use.}

\item{top_n}{Number of top variable feature to highlight. Default to 20.}

\item{cols}{2-value vectors that specify non-variable/variable status colors. Default to black for non-variable
and blue for variable.}
}
\value{
A ggplot object.
}
\description{
This function creates a dotplot of Average expression vs. Standardized Variance for each
feature, highlighting the most variable ones of a seurat object.
}
\examples{
data("pbmc_small")
plot_variable_features(pbmc_small)

# Specify the number of variable features to highlight
plot_variable_features(pbmc_small, top_n = 10)

}
\concept{DimRed}
