% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_violin_outliers}
\alias{plot_violin_outliers}
\title{Plot violin outlier thresholds}
\usage{
plot_violin_outliers(
  seurat_object = seurat_object,
  split_by = "orig.ident",
  metric = "nCount_RNA",
  title = "",
  x_trans = "identity",
  fill_vector = NULL
)
}
\arguments{
\item{seurat_object}{Seurat object to use.}

\item{split_by}{Name of the column used to group_by the violins (usually should correspond to the one used as \code{batch} in \code{\link[=calculate_qc_mad_outliers]{calculate_qc_mad_outliers()}}).
Default to "orig.ident".}

\item{metric}{Name of the column to plot. Default to "nCount_RNA".}

\item{title}{Title to add to the plot. Default to "".}

\item{x_trans}{Transformation to pass to scale_x_continuous(trans). Default to "identity".}

\item{fill_vector}{Optional named vector used to assign fill colors to the different group of fill variable.
Default to NULL.}
}
\value{
A ggplot list object with the plot.
}
\description{
This function creates is used to represent the distribution of the values of a desired metric (meta.data column)
and the corresponding thresholds (if available). It can help in the visual inspection of nmads to use when calculating
outliers with \code{\link[=calculate_qc_mad_outliers]{calculate_qc_mad_outliers()}}.
}
\examples{
data("pbmc_small")
pbmc_small <- calculate_qc_metrics(pbmc_small)
pbmc_small <- calculate_qc_mad_outliers(pbmc_small)

# Basic plot
plot_violin_outliers(pbmc_small)

# Changing metric
plot_violin_outliers(pbmc_small, metric = "riboRatio")

}
\concept{QC}
