% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc_correlation_plot}
\alias{qc_correlation_plot}
\title{Create a correlation plot between 2 different QC metrics.}
\usage{
qc_correlation_plot(
  seurat_object,
  metric_x = "nCount_RNA",
  metric_y = "nFeature_RNA",
  title = "",
  shape = "orig.ident",
  color = "mitoRatio",
  x_threshold = NULL,
  y_threshold = NULL,
  x_trans = "identity",
  y_trans = "identity"
)
}
\arguments{
\item{seurat_object}{Seurat object to use to create the plot.}

\item{metric_x}{Name of the metric (meta.data column) to plot on x axis. Default to "nCount_RNA".}

\item{metric_y}{Name of the metric (meta.data column) to plot on y axis. Default to "nFeature_RNA".}

\item{title}{Title to add to the plot. Default to "".}

\item{shape}{Name of the column to use to pass to aes(shape) in geom_point.
It is used to change the shape of the dots based on values of a meta.data column. Default to "orig.ident".}

\item{color}{Name of the column to use to pass to aes(color) in geom_point.
It is used to change the color of the dots based on values of a meta.data column. Default to "mitoRatio".}

\item{x_threshold}{Optional value of the threshold line to plot as a vertical line. Default to NULL, not plotted.}

\item{y_threshold}{Optional value of the threshold line to plot as an horizontal line. Default to NULL, not plotted.}

\item{x_trans}{Transformation to pass to scale_x_continuous(trans). Default to "identity".}

\item{y_trans}{Transformation to pass to scale_y_continuous(trans). Default to "identity".}
}
\value{
A ggplot list object with the plot.
}
\description{
This function creates a correlation plot between two given qc metric (or any continuous variable
of the Seurat object meta.data).
}
\examples{
data("pbmc_small")
pbmc_small <- calculate_qc_metrics(pbmc_small)

# Basic plot
qc_correlation_plot(pbmc_small)

# Specify x and y metrics
qc_correlation_plot(pbmc_small, metric_x = "mitoRatio",
                    metric_y = "riboRatio")

}
\concept{QC}
