% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc_density_plot}
\alias{qc_density_plot}
\title{Create a density plot for a QC metric}
\usage{
qc_density_plot(
  seurat_object,
  metric = "nCount_RNA",
  title = "nCount distribution",
  fill = "orig.ident",
  fill_vector = NULL,
  alpha = 0.2,
  threshold = NULL,
  x_trans = "identity"
)
}
\arguments{
\item{seurat_object}{Seurat object to use to create the plot.}

\item{metric}{Name of the metric (meta.data column) to plot. Default to "nCount_RNA".}

\item{title}{Title to add to the plot. Default to "nCount distribution".}

\item{fill}{Name of the column to use to pass to aes(fill) in geom_density.
It is used to split the data into groups and draw a density plot for each value of it. Default to "orig.ident".}

\item{fill_vector}{Optional named vector used to assign fill colors to the different group of fill variable.
Default to NULL.}

\item{alpha}{Value to pass to alpha in geom_density. Default to 0.2.}

\item{threshold}{Optional value of the threshold line to plot as a vertical line. Default to NULL, not plotted.}

\item{x_trans}{Transformation to pass to scale_x_continuous(trans). Default to "identity".}
}
\value{
A ggplot list object with the plot.
}
\description{
This function creates a density plot for the given qc metric (or any continuous variable
of the Seurat object meta.data).
}
\examples{
data("pbmc_small")
pbmc_small <- calculate_qc_metrics(pbmc_small)

# Basic plot
qc_density_plot(pbmc_small)

# Specify fill colors
qc_density_plot(pbmc_small, fill = "groups", fill_vector = c("g1" = "blue", "g2" = "red"))

}
\concept{QC}
