% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schelp-package.R
\docType{package}
\name{schelp-package}
\alias{schelp}
\alias{schelp-package}
\title{schelp: Data processing and analysis tools for scRNAseq data}
\description{
A collection of helper functions that I use for my scRNAseq data processing and analysis pipelines.
}
\author{
\strong{Maintainer}: Victor Yuan \email{victor.2wy@gmail.com}

}
\keyword{internal}
