% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_qc.R
\name{summarize_qc}
\alias{summarize_qc}
\title{Summarize quality control data}
\usage{
summarize_qc(df, ngene_th = 250, numi_th = 500, mtr_th = 0.2)
}
\arguments{
\item{df}{a tibble or data frame with the following columns:
n_gene, n_umi, qc_mtfraction, scrublet_call, qc_log10genesperumi}

\item{ngene_th}{a numeric that is the lower bound of number of genes detected
per cell}

\item{numi_th}{a numeric that is the lower bound of the number of UMIs
detected per cell}

\item{mtr_th}{ranges between 0-1. upper bound of the fraction of UMIs mapping
to mitochondrial genes}
}
\value{
a tibble containing summarized qc info
}
\description{
Calculates statistics across cells for # genes detected, # UMIs, and
mitochondrial gene fraction.
}
\details{
Will also count the number of cells failing QC thresholds.

Pass a grouped tibble to calculate statistics by group.
}
