% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{view_default_qc_mads}
\alias{view_default_qc_mads}
\title{View default QC MAD parameters}
\usage{
view_default_qc_mads()
}
\value{
A tibble with columns:
\itemize{
\item metric: name of the metric
\item lower.nmads: value of the lower nmads used to calculate outliers
\item higher.nmads: value of the higher nmads used to calculate outliers
\item log: whether outlier detection is done on log-transformed data
}
}
\description{
This function is used to see the default values for \code{\link[=calculate_qc_mad_outliers]{calculate_qc_mad_outliers()}} metrics.
}
\examples{
view_default_qc_mads()

}
\concept{QC}
