% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{calculate_qc_metrics}
\alias{calculate_qc_metrics}
\title{Calculate QC metrics}
\usage{
calculate_qc_metrics(
  seurat_object,
  assay = NULL,
  mito = "^MT-",
  ribo = "^RP[SL]",
  hemo = "^HB[^(P)]"
)
}
\arguments{
\item{seurat_object}{Seurat object of which calculate QC metrics.}

\item{assay}{Assay to use. By default it takes the current active assay.}

\item{mito}{Pattern to use to match for mitochondrial genes. Default to "^MT-".}

\item{ribo}{Pattern to use to match for ribosomal genes. Default to "^RP[SL]".}

\item{hemo}{Pattern to use to match for hemoglobin genes. Default to "^HB[^(P)]".}
}
\value{
A Seurat object with 4 new columns in meta.data:
\itemize{
\item log10GenesPerUMI
\item mitoRatio
\item riboRatio
\item hemoRatio
}
}
\description{
This function calculates QC metrics (log10GenesPerUMI, mitochondrial gene ratio, ribosomal gene ratio and hemoglobin gene ratio)
for the provided seurat object.
}
\examples{
data("pbmc_small")
pbmc_small <- calculate_qc_metrics(pbmc_small)

}
\concept{QC}
