% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_qc_metrics.R
\name{calculate_qc_metrics}
\alias{calculate_qc_metrics}
\title{calculate_qc_metrics}
\usage{
calculate_qc_metrics(scobj)
}
\arguments{
\item{scobj}{a seurat object}
}
\value{
a tibble
}
\description{
calculate_qc_metrics
}
