% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gene.R
\name{extract_gene}
\alias{extract_gene}
\title{extracts gene expression data into metadata}
\usage{
extract_gene(metadata, cellid_col = "cellid", expr, genes)
}
\arguments{
\item{metadata}{cell metadata}

\item{cellid_col}{column in metadata containing cell identifiers, should
match rownames(counts). Default is "cellid"}

\item{expr}{count matrix, cells in columns, genes in rows}

\item{genes}{character vector of gene names}
}
\value{
tibble
}
\description{
First checks if gene names are in counts rownames
If genes already exist in metadata, message that they are being overwritten
rownames of counts will be joined on a cell identifier column in metadata
using left_join
}
