% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_cellgene.R
\name{plot_bar_cellgene}
\alias{plot_bar_cellgene}
\title{plots expression by cell cluster}
\usage{
plot_bar_cellgene(
  metadata,
  genes,
  celltype,
  type = "boxplot",
  zscore = FALSE,
  add_jitter = FALSE,
  jitterheight = 0.1,
  alpha = 1,
  pointsize = 1,
  ncol = NULL,
  nrow = NULL
)
}
\arguments{
\item{metadata}{cells metadata}

\item{genes}{character vector of gene names}

\item{celltype}{column of metadata that denotes cell identity}

\item{zscore}{Whether expression should be zscore or not}
}
\value{
ggplot2
}
\description{
Use extract_gene so that gene expression exists as columns of metadata.
}
