% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_cellgene.R
\name{plot_bar_cellgene}
\alias{plot_bar_cellgene}
\title{plots expression by cell cluster}
\usage{
plot_bar_cellgene(
  metadata,
  cellid_col = "cellid",
  expr,
  genes,
  celltype,
  type = "boxplot",
  zscore = FALSE,
  add_jitter = FALSE,
  jitterheight = 0.75,
  alpha = 0.5,
  pointsize = 1,
  ncol = NULL,
  nrow = NULL,
  fontsize_p = 1
)
}
\arguments{
\item{metadata}{cell metadata}

\item{cellid_col}{column in metadata containing cell identifiers, should
match rownames(counts). Default is "cellid"}

\item{expr}{count matrix, cells in columns, genes in rows}

\item{genes}{character vector of gene names}

\item{type}{one of "mean_se", "bar_mean", "bar_mean_se", "bar_median",
"violin", "boxplot"}

\item{zscore}{Whether expression should be zscore or not}

\item{add_jitter}{adds jittered points default = FALSE,}

\item{jitterheight}{multiplier that controls jitter height. Default = 0.75}

\item{alpha}{controls the alpha of jittered points = 0.5}

\item{pointsize}{= 1,}

\item{ncol}{for facetwrap}

\item{nrow}{for facetwrap}

\item{fontsize_p}{Multiplier that controls relative fontsize. Default is 1}
}
\value{
ggplot2
}
\description{
Use extract_gene so that gene expression exists as columns of metadata.
}
