% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schelp-package.R
\docType{package}
\name{schelp-package}
\alias{schelp}
\alias{schelp-package}
\title{schelp: Single Cell Helper Functions}
\description{
This package contains helper functions for the various steps of Single-Cell analysis (QC, dimensionality reduction, clustering, etc.).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mmiots9/schelp}
  \item Report bugs at \url{https://github.com/mmiots9/schelp/issues}
}

}
\author{
\strong{Maintainer}: Matteo Miotto \email{miotsdata@gmail.com} (\href{https://orcid.org/0000-0003-3288-9011}{ORCID})

}
\keyword{internal}
